Refactor file structure and prettify code
