+++ client/public/assets/favicon/android-chrome-192x192.png
