+++ client/public/assets/favicon/favicon-32x32.png
