+++ client/public/assets/favicon/favicon.ico
