+++ client/public/index.html
@@ -5,9 +5,9 @@
   <meta charset="utf-8" />
   <!-- Favicon -->
 
-  <link rel="apple-touch-icon" sizes="180x180" href="%PUBLIC_URL%/favicon/apple-touch-icon.png">
-  <link rel="icon" type="image/png" sizes="32x32" href="%PUBLIC_URL%/favicon/favicon-32x32.png">
-  <link rel="icon" type="image/png" sizes="16x16" href="%PUBLIC_URL%/favicon/favicon-16x16.png">
+  <link href="%PUBLIC_URL%/assets/favicon/apple-touch-icon.png" rel="apple-touch-icon" sizes="180x180">
+  <link href="%PUBLIC_URL%/assets/favicon/favicon-32x32.png" rel="icon" sizes="32x32" type="image/png">
+  <link href="%PUBLIC_URL%/assets/favicon/favicon-16x16.png" rel="icon" sizes="16x16" type="image/png">
   <link rel="manifest" href="/site.webmanifest">
 
   <meta name="theme-color" content="#000000" />
@@ -17,9 +17,6 @@
   <!-- Using Google Font -->
   <link rel="preconnect" href="https://fonts.gstatic.com">
   <link href="https://fonts.googleapis.com/css2?family=Public+Sans:wght@400;500;600;700&display=swap" rel="stylesheet">
-
-  <!-- Using Local Font -->
-  <link rel="stylesheet" type="text/css" href="%PUBLIC_URL%/fonts/index.css" />
 
   <!-- Css simplebar -->
   <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/simplebar/5.3.9/simplebar.min.css" />
