+++ client/src/constants.js
@@ -1,0 +1,22 @@
+const backendApiUrl = "http://localhost:8080/api";
+
+const routes = {
+  AUTHOR: "author",
+  AUTH: "auth",
+  BOOK: "book",
+  BORROWAL: "borrowal",
+  GENRE: "genre",
+  USER: "user"
+};
+
+const methods = {
+  GET: "get",
+  GET_ALL: "getAll",
+  POST: "add",
+  PUT: "update",
+  DELETE: "delete"
+};
+
+const apiUrl = (route, method, id = "") => `${backendApiUrl}/${route}/${method}${id && `/${id}`}`;
+
+module.exports = { routes, methods, apiUrl };
