+++ client/src/hooks/useLocalStorage.js
@@ -1,4 +1,4 @@
-import {useState} from "react";
+import { useState } from "react";
 
 export const useLocalStorage = (keyName, defaultValue) => {
   const [storedValue, setStoredValue] = useState(() => {
