+++ client/src/index.js
@@ -1,18 +1,18 @@
-import ReactDOM from 'react-dom/client';
-import {BrowserRouter} from 'react-router-dom';
-import {HelmetProvider} from 'react-helmet-async';
-import App from './App';
-import * as serviceWorker from './serviceWorker';
-import reportWebVitals from './reportWebVitals';
-import {AuthProvider} from "./useAuth";
+import ReactDOM from "react-dom/client";
+import { BrowserRouter } from "react-router-dom";
+import { HelmetProvider } from "react-helmet-async";
+import App from "./App";
+import * as serviceWorker from "./serviceWorker";
+import reportWebVitals from "./reportWebVitals";
+import { AuthProvider } from "./hooks/useAuth";
 
-const root = ReactDOM.createRoot(document.getElementById('root'));
+const root = ReactDOM.createRoot(document.getElementById("root"));
 
 root.render(
   <HelmetProvider>
     <BrowserRouter>
       <AuthProvider>
-        <App/>
+        <App />
       </AuthProvider>
     </BrowserRouter>
   </HelmetProvider>
