+++ client/src/layouts/dashboard/LibraryApp.jsx
@@ -5,7 +5,7 @@ import { styled } from "@mui/material/styles";
 //
 import Header from "./header";
 import Nav from "./nav";
-import { useAuth } from "../../useAuth";
+import { useAuth } from "../../hooks/useAuth";
 
 // ----------------------------------------------------------------------
