+++ client/src/layouts/dashboard/nav/index.js
@@ -13,7 +13,7 @@ import Scrollbar from "../../../components/scrollbar";
 import NavSection from "../../../components/nav-section";
 //
 import navConfig from "./config";
-import { useAuth } from "../../../useAuth";
+import { useAuth } from "../../../hooks/useAuth";
 
 // ----------------------------------------------------------------------
