+++ client/src/routes.js
@@ -1,32 +1,32 @@
 import { Navigate, useRoutes } from "react-router-dom";
 import LibraryApp from "./layouts/dashboard";
-import AuthorPage from "./pages/AuthorPage";
-import LoginPage from "./pages/LoginPage";
+import AuthorPage from "./sections/@dashboard/author/AuthorPage";
+import LoginPage from "./sections/auth/login/LoginPage";
 import Page404 from "./pages/Page404";
-import BorrowalPage from "./pages/BorrowalPage";
-import BookPage from "./pages/BookPage";
-import DashboardAppPage from "./pages/DashboardAppPage";
-import UsersPage from "./pages/UserPage";
-import GenrePage from "./pages/GenrePage";
-import { useAuth } from "./useAuth";
+import BorrowalPage from "./sections/@dashboard/borrowal/BorrowalPage";
+import BookPage from "./sections/@dashboard/book/BookPage";
+import DashboardAppPage from "./sections/@dashboard/app/DashboardAppPage";
+import UsersPage from "./sections/@dashboard/user/UserPage";
+import GenrePage from "./sections/@dashboard/genre/GenrePage";
+import { useAuth } from "./hooks/useAuth";
 
 // ----------------------------------------------------------------------
 
 export default function Router() {
-  const {user} = useAuth();
+  const { user } = useAuth();
   const adminRoutes = useRoutes([
     {
-      path: '/',
-      element: <LibraryApp/>,
+      path: "/",
+      element: <LibraryApp />,
       children: [
-        {element: <Navigate to="/app"/>, index: true},
-        {path: 'dashboard', element: <DashboardAppPage/>},
-        {path: 'authors', element: <AuthorPage/>},
-        {path: 'books', element: <BookPage/>},
-        {path: 'borrowals', element: <BorrowalPage/>},
-        {path: 'genres', element: <GenrePage/>},
-        {path: 'users', element: <UsersPage/>},
-      ],
+        { element: <Navigate to="/app" />, index: true },
+        { path: "dashboard", element: <DashboardAppPage /> },
+        { path: "authors", element: <AuthorPage /> },
+        { path: "books", element: <BookPage /> },
+        { path: "borrowals", element: <BorrowalPage /> },
+        { path: "genres", element: <GenrePage /> },
+        { path: "users", element: <UsersPage /> }
+      ]
     },
     {
       path: 'login',
